requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
plotly==5.18.0
scikit-learn==1.3.2
python-dotenv==1.0.0
